fix type

+++ strategies/multi_coin_abnormal_volumes_tracker/models/pair_manager_dict.py
@@ -12,4 +12,4 @@ class PairManagerDict(TypedDict):
     position_driver: Optional[PositionDriver]
     last_position_driver_state: PositionStateEnum
     jail_start_timestamp: int
-    invested: int
+    invested: float
